Revert "Kleine push voor difficulty moet nog iets beter worden uitgewerkt."
This reverts commit 0da44aebab2e72f4923968b56ebd269bb4f92d69.

+++ CustomPlatformerController/Assets/Scripts/Base/Stats.cs
@@ -9,7 +9,7 @@ namespace Assets.Scripts.Base
         public int handle;
         public int fight;
         public int hide;
-        public float levelDifficulty;
+        public int levelDifficulty;
         public int coins;
         public int amountOfPlaythroughs;
         public DateTime lastFinishedPlaythrough;
